loss function modularization
